Move explicit export list to Foreign.JNI.Types.
This allows exporting the Sing data instances explicitly, which is the
only way I know to export it at all. Otherwise the data instance
consructors are not in scope in importing modules.
